5-flip_bits.c

+++ 0x14-bit_manipulation/5-flip_bits.c
@@ -12,7 +12,7 @@
 
 unsigned int flip_bits(unsigned long int n, unsigned long int m)
 {
-	unsigned int counter = 0;
+	unsigned long int counter = 0;
 	unsigned int diff = n ^ m;/* XOR operation to get the differing bits */
 
 	while (diff != 0)
